refactor(widget): move protocol ownership to widget-js
