+++ .github/workflows/widget-cdn-deploy.yml
@@ -185,8 +185,8 @@ jobs:
       - name: Install dependencies
         run: pnpm install --frozen-lockfile
 
-      - name: Build shared widget protocol
-        run: pnpm --filter @layerswap/widget-types build
+      - name: Build widget loader protocol
+        run: pnpm --filter @layerswap/widget-js... build
 
       - name: Download approved release candidate
         uses: actions/download-artifact@d3f86a106a0bac45b974a628896c90dbdf5c8093 # v4
