+++ apps/widget-cdn/scripts/protocol-contract.test.mjs
@@ -3,7 +3,7 @@ import { readFileSync } from 'node:fs';
 import { dirname, join, resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import test from 'node:test';
-import { WIDGET_PROTOCOL_MAJOR } from '@layerswap/widget-types';
+import { WIDGET_PROTOCOL_MAJOR } from '@layerswap/widget-js';
 
 const root = resolve(dirname(fileURLToPath(import.meta.url)), '../../..');
 
@@ -13,7 +13,6 @@ function packageMajor(relativePath) {
 }
 
 test('public loader package majors match the CDN protocol major', () => {
-    assert.equal(packageMajor('packages/widget/types/package.json'), WIDGET_PROTOCOL_MAJOR);
     assert.equal(packageMajor('packages/widget/js/package.json'), WIDGET_PROTOCOL_MAJOR);
     assert.equal(packageMajor('packages/widget/react/package.json'), WIDGET_PROTOCOL_MAJOR);
 });
