+++ packages/widget/types/src/index.ts
@@ -1,5 +1,4 @@
 export type { ThemeData, ThemeColor, StatusColor } from './theme';
-export { WIDGET_PROTOCOL_MAJOR, widgetProtocolMajorOf } from './protocol.js';
 export type {
   WidgetConfig,
   WidgetCallbacks,
